Fix Skip Connections, Add distance plots, add PID check
The calculation of the chain length for Skip Connection plots was changed as the original method caused issues at the Si-Scintillator boundaries.

Several distance plots in x-y and eta-phi were added based on the position gained through the recHitTools_ using detids.

A check was added to find the particle ids found in the events.